Fix missing git status information in the working tree root.
New git now fails on empty pathspecs with this message:
fatal: empty string is not a valid pathspec. please use . instead if you meant to match all paths

By making $exclude an array, it now expands to nothing or the exclude
pattern, instead of an empty string or the exclude pattern.